Split frontend API facade helpers
