+++ frontend/lib/api.ts
@@ -1,16 +1,3 @@
-import { useEffect, useRef, useState } from 'react';
-import useSWR from 'swr';
-import useSWRInfinite from 'swr/infinite';
-import { authFetch } from '@/lib/authFetch';
-import { readLastKnownUserId, writeLastKnownUserId } from '@/lib/last-known';
-import { readBrowserSession } from '@/lib/supabase-auth-cleanup';
-import { hasSupabaseAuthCookie } from '@/lib/supabase-session-hint';
-import type { EnginesResponse } from '@/types/engines';
-import type { Job, JobsPage } from '@/types/jobs';
-import type { PricingSnapshot } from '@maxvideoai/pricing';
-import { normalizeJobMessage, normalizeJobProgress, normalizeJobStatus } from '@/lib/job-status';
-import type { JobSurface } from '@/types/billing';
-
 export {
   runAngleTool,
   runAudioGenerate,
@@ -21,686 +8,7 @@ export {
   runUpscaleTool,
 } from '@/lib/api-generation';
 
-async function getSupabaseClient() {
-  const { supabase } = await import('@/lib/supabaseClient');
-  return supabase;
-}
-
-type JobStatusResult = {
-  ok: true;
-  jobId: string;
-  status: 'pending' | 'completed' | 'failed';
-  progress: number;
-  videoUrl: string | null;
-  previewVideoUrl?: string | null;
-  audioUrl?: string | null;
-  thumbUrl: string | null;
-  aspectRatio?: string | null;
-  pricing?: PricingSnapshot;
-  finalPriceCents?: number;
-  currency?: string;
-  paymentStatus: string;
-  batchId?: string | null;
-  groupId?: string | null;
-  iterationIndex?: number | null;
-  iterationCount?: number | null;
-  renderIds?: string[] | null;
-  renderThumbUrls?: string[] | null;
-  heroRenderId?: string | null;
-  localKey?: string | null;
-  message?: string | null;
-  etaSeconds?: number | null;
-  etaLabel?: string | null;
-};
-
-type StatusRetryMeta = {
-  attempt: number;
-  timer: number | null;
-};
-
-function jobHasRenderableMedia(job: Pick<Job, 'videoUrl' | 'audioUrl' | 'renderIds'>): boolean {
-  const hasImageMedia =
-    Array.isArray(job.renderIds) && job.renderIds.some((value) => typeof value === 'string' && value.length);
-  return Boolean(job.videoUrl || job.audioUrl) || hasImageMedia;
-}
-
-function normalizeJobFromApi(job: Job): Job {
-  const hasImageMedia =
-    Array.isArray(job.renderIds) && job.renderIds.some((value) => typeof value === 'string' && value.length);
-  const hasMedia = Boolean(job.videoUrl || job.audioUrl) || hasImageMedia;
-  const normalizedStatus = normalizeJobStatus(job.status ?? null, hasMedia);
-  const status: Job['status'] =
-    normalizedStatus ?? (hasMedia ? 'completed' : 'pending');
-  const progress = normalizeJobProgress(job.progress, status as 'pending' | 'completed' | 'failed', hasMedia);
-  const messageFromApi = normalizeJobMessage(job.message);
-  const message =
-    hasImageMedia && status === 'completed'
-      ? undefined
-      : messageFromApi ??
-        (status === 'failed'
-          ? 'The service reported a failure without details. Try again. If it fails repeatedly, contact support with your request ID.'
-          : undefined);
-
-  return {
-    ...job,
-    status,
-    progress,
-    message,
-  };
-}
-
-const STATUS_RETRY_TIMERS = new Map<string, StatusRetryMeta>();
-const STATUS_RETRY_BASE_DELAY_MS = 60_000;
-const STATUS_RETRY_MAX_DELAY_MS = 30 * 60 * 1000;
-
-function clearStatusRetry(jobId: string): void {
-  if (typeof window === 'undefined') return;
-  const existing = STATUS_RETRY_TIMERS.get(jobId);
-  if (existing?.timer != null) {
-    window.clearTimeout(existing.timer);
-  }
-  STATUS_RETRY_TIMERS.delete(jobId);
-}
-
-function scheduleStatusRetry(jobId: string, attempt: number): void {
-  if (typeof window === 'undefined') return;
-  const existing = STATUS_RETRY_TIMERS.get(jobId);
-  if (existing?.timer != null) {
-    window.clearTimeout(existing.timer);
-  }
-  const delay = Math.min(STATUS_RETRY_BASE_DELAY_MS * Math.pow(2, attempt), STATUS_RETRY_MAX_DELAY_MS);
-  const timer = window.setTimeout(() => {
-    STATUS_RETRY_TIMERS.set(jobId, { attempt: attempt + 1, timer: null });
-    void getJobStatus(jobId).catch(() => {
-      scheduleStatusRetry(jobId, attempt + 1);
-    });
-  }, delay);
-  STATUS_RETRY_TIMERS.set(jobId, { attempt, timer });
-}
-
-type EngineCategory = 'video' | 'image' | 'all';
-
-type UseEnginesOptions = {
-  includeAverages?: boolean;
-  enabled?: boolean;
-};
-
-async function loadFallbackEngines(category: EngineCategory): Promise<EnginesResponse['engines']> {
-  const { getBaseEnginesByCategory } = await import('@/lib/engines');
-  return getBaseEnginesByCategory(category);
-}
-
-export function useEngines(category: EngineCategory = 'video', options?: UseEnginesOptions) {
-  const enabled = options?.enabled !== false;
-  const params = new URLSearchParams();
-  if (category !== 'video') {
-    params.set('category', category);
-  }
-  if (options?.includeAverages) {
-    params.set('includeAverages', '1');
-  }
-  const query = params.size > 0 ? `?${params.toString()}` : '';
-  return useSWR<EnginesResponse>(
-    enabled ? `static-engines:${category}:${options?.includeAverages ? 'avg' : 'base'}` : null,
-    async () => {
-      try {
-        const response = await fetch(`/api/engines${query}`, { credentials: 'include' });
-        const data = (await response.json().catch(() => null)) as
-          | { engines?: EnginesResponse['engines']; error?: string }
-          | null;
-        if (!response.ok) {
-          throw new Error(data?.error ?? `Engines request failed: ${response.status}`);
-        }
-        return { engines: data?.engines ?? [] };
-      } catch {
-        const fallbackEngines = await loadFallbackEngines(category);
-        return { engines: fallbackEngines };
-      }
-    },
-    {
-      dedupingInterval: 5 * 60 * 1000,
-    }
-  );
-}
-
-type JobFeedType = 'video' | 'image' | 'all';
-type JobFeedSurface = JobSurface | 'all';
-
-async function fetchJobsPage(
-  limit: number,
-  cursor?: string | null,
-  options?: { type?: JobFeedType; surface?: JobFeedSurface }
-): Promise<JobsPage> {
-  const params = new URLSearchParams({ limit: String(limit) });
-  if (cursor) params.set('cursor', cursor);
-  if (options?.surface && options.surface !== 'all') {
-    params.set('surface', options.surface);
-  } else if (options?.type && options.type !== 'all') {
-    params.set('type', options.type);
-  }
-  const response = await authFetch(`/api/jobs?${params.toString()}`);
-
-  const payload = (await response.json().catch(() => null)) as (Partial<JobsPage> & { error?: string }) | null;
-
-  if (!response.ok) {
-    const message = payload?.error ?? `Jobs request failed: ${response.status}`;
-    throw new Error(message);
-  }
-
-  if (!payload || !Array.isArray(payload.jobs)) {
-    throw new Error('Jobs payload malformed');
-  }
-
-  const jobs = payload.jobs.map((job) => normalizeJobFromApi(job));
-
-  return {
-    ok: payload.ok !== false,
-    jobs,
-    nextCursor: payload.nextCursor ?? null,
-  };
-}
-
-type JobsKey = readonly ['jobs', string, number, string | null, JobFeedType, JobFeedSurface];
-
-export function useInfiniteJobs(pageSize = 12, options?: { type?: JobFeedType; surface?: JobFeedSurface }) {
-  const [cacheKey, setCacheKey] = useState<string | null>(() =>
-    typeof window === 'undefined' ? null : readLastKnownUserId()
-  );
-  const feedType: JobFeedType =
-    options?.type === 'image' || options?.type === 'video' ? options.type : 'all';
-  const feedSurface: JobFeedSurface =
-    options?.surface === 'video' ||
-    options?.surface === 'image' ||
-    options?.surface === 'audio' ||
-    options?.surface === 'character' ||
-    options?.surface === 'angle' ||
-    options?.surface === 'upscale'
-      ? options.surface
-      : 'all';
-  const lastRevalidateRef = useRef<number>(0);
-  const lastKnownUserIdRef = useRef<string | null>(typeof window === 'undefined' ? null : readLastKnownUserId());
-  const [stableStore, setStableStore] = useState<{
-    byId: Record<string, Job>;
-    order: string[];
-  }>({ byId: {}, order: [] });
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return undefined;
-    let cancelled = false;
-    const updateCacheKey = async (forceAnonymous = false) => {
-      if (forceAnonymous) {
-        if (cancelled) return;
-        lastKnownUserIdRef.current = null;
-        writeLastKnownUserId(null);
-        setCacheKey('anonymous');
-        return;
-      }
-      const knownUserId = lastKnownUserIdRef.current ?? readLastKnownUserId();
-      if (!knownUserId && !hasSupabaseAuthCookie()) {
-        setCacheKey('anonymous');
-        return;
-      }
-      const session = await readBrowserSession();
-      if (cancelled) return;
-      const userId = session?.user?.id ?? null;
-      if (userId) {
-        lastKnownUserIdRef.current = userId;
-        writeLastKnownUserId(userId);
-        setCacheKey(userId);
-        return;
-      }
-      if (lastKnownUserIdRef.current) {
-        setCacheKey(lastKnownUserIdRef.current);
-        return;
-      }
-      const storedUserId = readLastKnownUserId();
-      if (storedUserId) {
-        lastKnownUserIdRef.current = storedUserId;
-        setCacheKey(storedUserId);
-        return;
-      }
-      setCacheKey('anonymous');
-    };
-    void updateCacheKey();
-    let subscription: { data?: { subscription?: { unsubscribe: () => void } } } | null = null;
-    if (lastKnownUserIdRef.current || readLastKnownUserId() || hasSupabaseAuthCookie()) {
-      void getSupabaseClient().then((supabase) => {
-        if (cancelled) return;
-        subscription = supabase.auth.onAuthStateChange(async (event, session) => {
-          if (cancelled) return;
-          const eventType = event as string;
-          if (eventType === 'SIGNED_OUT' || eventType === 'USER_DELETED') {
-            await updateCacheKey(true);
-            return;
-          }
-          const nextUserId = session?.user?.id ?? null;
-          if (nextUserId) {
-            lastKnownUserIdRef.current = nextUserId;
-            writeLastKnownUserId(nextUserId);
-            setCacheKey(nextUserId);
-            return;
-          }
-          await updateCacheKey();
-        });
-      });
-    }
-    return () => {
-      cancelled = true;
-      subscription?.data?.subscription?.unsubscribe();
-    };
-  }, []);
-
-  useEffect(() => {
-    setStableStore({ byId: {}, order: [] });
-  }, [cacheKey, feedSurface, feedType]);
-
-  const getJobsKey = (index: number, previousPage: JobsPage | null | undefined): JobsKey | null => {
-    if (!cacheKey) return null;
-    if (previousPage && !previousPage.nextCursor) {
-      return null;
-    }
-    const cursor = index === 0 ? null : previousPage?.nextCursor ?? null;
-    return ['jobs', cacheKey, pageSize, cursor, feedType, feedSurface];
-  };
-
-  const fetchJobs = async (key: JobsKey) => {
-    const [, , limit, cursor, type, surface] = key;
-    return fetchJobsPage(limit, cursor, { type, surface });
-  };
-
-  const swr = useSWRInfinite<JobsPage, Error>(getJobsKey, fetchJobs);
-
-  const { mutate } = swr;
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return undefined;
-    const handler = (event: Event) => {
-      const detail = (event as CustomEvent<JobStatusResult>).detail;
-      if (!detail || !detail.jobId) return;
-
-      const normalizedStatus = typeof detail.status === 'string' ? detail.status.toLowerCase() : '';
-      if (normalizedStatus === 'completed' || normalizedStatus === 'failed') {
-        clearStatusRetry(detail.jobId);
-      } else {
-        const meta = STATUS_RETRY_TIMERS.get(detail.jobId);
-        if (!meta || meta.timer === null) {
-          scheduleStatusRetry(detail.jobId, meta?.attempt ?? 0);
-        }
-      }
-
-      let jobFound = false;
-      void mutate(
-        (pages) => {
-          if (!pages) return pages;
-          const nextPages = pages.map((page) => {
-            let pageModified = false;
-            const jobs = page.jobs.map((job) => {
-              if (job.jobId !== detail.jobId) {
-                return job;
-              }
-              jobFound = true;
-              pageModified = true;
-              const progressFromDetail =
-                typeof detail.progress === 'number' && Number.isFinite(detail.progress)
-                  ? Math.max(0, Math.min(100, detail.progress))
-                  : undefined;
-              const next = {
-                ...job,
-                status: detail.status ?? job.status,
-                progress:
-                  typeof progressFromDetail === 'number' && progressFromDetail > 0
-                    ? progressFromDetail
-                    : job.progress,
-                videoUrl: detail.videoUrl ?? job.videoUrl,
-                previewVideoUrl: detail.previewVideoUrl ?? job.previewVideoUrl,
-                audioUrl: detail.audioUrl ?? job.audioUrl,
-                thumbUrl: detail.thumbUrl ?? job.thumbUrl,
-                finalPriceCents: detail.finalPriceCents ?? job.finalPriceCents,
-                currency: detail.currency ?? job.currency,
-                pricingSnapshot: detail.pricing ?? job.pricingSnapshot,
-                paymentStatus: detail.paymentStatus ?? job.paymentStatus,
-                batchId: detail.batchId ?? job.batchId,
-                groupId: detail.groupId ?? job.groupId,
-                iterationIndex: detail.iterationIndex ?? job.iterationIndex,
-                iterationCount: detail.iterationCount ?? job.iterationCount,
-                heroRenderId: detail.heroRenderId ?? job.heroRenderId,
-                localKey: detail.localKey ?? job.localKey,
-                message: detail.message ?? job.message,
-                etaSeconds: detail.etaSeconds ?? job.etaSeconds,
-                etaLabel: detail.etaLabel ?? job.etaLabel,
-              };
-              if (detail.renderIds !== undefined) {
-                next.renderIds = detail.renderIds;
-              }
-              if (detail.renderThumbUrls !== undefined) {
-                next.renderThumbUrls = detail.renderThumbUrls;
-              }
-              return next;
-            });
-            return pageModified ? { ...page, jobs } : page;
-          });
-          return jobFound ? nextPages : pages;
-        },
-        { revalidate: false }
-      );
-
-      if (!jobFound) {
-        const now = Date.now();
-        if (now - lastRevalidateRef.current > 1500) {
-          lastRevalidateRef.current = now;
-          void mutate(undefined, { revalidate: true });
-        }
-      }
-    };
-
-    const errorHandler = (event: Event) => {
-      const detail = (event as CustomEvent<{ jobId?: string }>).detail;
-      const jobId = detail?.jobId;
-      if (!jobId) return;
-      const meta = STATUS_RETRY_TIMERS.get(jobId);
-      if (meta?.timer) return;
-      scheduleStatusRetry(jobId, meta?.attempt ?? 0);
-      void mutate(undefined, { revalidate: true });
-    };
-
-    window.addEventListener('jobs:status', handler as EventListener);
-    window.addEventListener('jobs:status-error', errorHandler as EventListener);
-    return () => {
-      window.removeEventListener('jobs:status', handler as EventListener);
-      window.removeEventListener('jobs:status-error', errorHandler as EventListener);
-    };
-  }, [mutate]);
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return undefined;
-    const handler = (event: Event) => {
-      const detail = (event as CustomEvent<{ jobId?: string }>).detail;
-      const jobId = typeof detail?.jobId === 'string' ? detail.jobId : '';
-      if (!jobId) return;
-      setStableStore((prev) => {
-        if (!prev.byId[jobId]) return prev;
-        const rest = { ...prev.byId };
-        delete rest[jobId];
-        const order = prev.order.filter((id) => id !== jobId);
-        return { byId: rest, order };
-      });
-      void mutate(
-        (pages) => {
-          if (!pages) return pages;
-          return pages.map((page) => ({ ...page, jobs: page.jobs.filter((job) => job.jobId !== jobId) }));
-        },
-        { revalidate: false }
-      );
-    };
-    window.addEventListener('jobs:hidden', handler as EventListener);
-    return () => window.removeEventListener('jobs:hidden', handler as EventListener);
-  }, [mutate]);
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-    const jobs = swr.data?.flatMap((page) => page.jobs) ?? [];
-    if (!jobs.length) {
-      if (swr.data) {
-        setStableStore({ byId: {}, order: [] });
-      }
-      return;
-    }
-
-    setStableStore((prev) => {
-      const currentJobIds = new Set(
-        jobs
-          .map((job) => (typeof job?.jobId === 'string' ? job.jobId : ''))
-          .filter((jobId): jobId is string => jobId.length > 0)
-      );
-      const byId: Record<string, Job> = {};
-      Object.entries(prev.byId).forEach(([jobId, job]) => {
-        if (currentJobIds.has(jobId)) {
-          byId[jobId] = job;
-        }
-      });
-      let sawRealJob = false;
-      jobs.forEach((job) => {
-        if (!job?.jobId) return;
-        if (!job.curated) {
-          sawRealJob = true;
-        }
-        const existing = byId[job.jobId];
-        if (!existing) {
-          byId[job.jobId] = job;
-          return;
-        }
-
-        const merged: Job = { ...existing, ...job };
-        if (job.thumbUrl == null && existing.thumbUrl != null) merged.thumbUrl = existing.thumbUrl;
-        if (job.videoUrl == null && existing.videoUrl != null) merged.videoUrl = existing.videoUrl;
-        if (job.audioUrl == null && existing.audioUrl != null) merged.audioUrl = existing.audioUrl;
-        if (job.readyVideoUrl == null && existing.readyVideoUrl != null) merged.readyVideoUrl = existing.readyVideoUrl;
-        if (job.previewFrame == null && existing.previewFrame != null) merged.previewFrame = existing.previewFrame;
-        if (job.renderIds == null && existing.renderIds != null) merged.renderIds = existing.renderIds;
-        if (job.renderThumbUrls == null && existing.renderThumbUrls != null) {
-          merged.renderThumbUrls = existing.renderThumbUrls;
-        }
-        if (job.heroRenderId == null && existing.heroRenderId != null) merged.heroRenderId = existing.heroRenderId;
-        byId[job.jobId] = merged;
-      });
-
-      const ids = Object.keys(byId);
-      if (sawRealJob) {
-        ids.forEach((id) => {
-          if (byId[id]?.curated) {
-            delete byId[id];
-          }
-        });
-      }
-
-      const MAX_STABLE_JOBS = 400;
-      const nextOrder = Object.keys(byId)
-        .sort((a, b) => {
-          const timeA = Date.parse(byId[a]?.createdAt ?? '');
-          const timeB = Date.parse(byId[b]?.createdAt ?? '');
-          return (Number.isFinite(timeB) ? timeB : 0) - (Number.isFinite(timeA) ? timeA : 0);
-        })
-        .slice(0, MAX_STABLE_JOBS);
-      if (nextOrder.length < Object.keys(byId).length) {
-        const keep = new Set(nextOrder);
-        Object.keys(byId).forEach((id) => {
-          if (!keep.has(id)) delete byId[id];
-        });
-      }
-
-      return { byId, order: nextOrder };
-    });
-  }, [swr.data]);
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return;
-    const jobs = swr.data?.flatMap((page) => page.jobs) ?? [];
-    const seen = new Set<string>();
-    jobs.forEach((job) => {
-      const jobId = job?.jobId;
-      if (!jobId) return;
-      seen.add(jobId);
-      const normalizedStatus =
-        normalizeJobStatus(job.status ?? null, jobHasRenderableMedia(job)) ??
-        (jobHasRenderableMedia(job) ? 'completed' : 'pending');
-      if (normalizedStatus === 'completed' || normalizedStatus === 'failed') {
-        clearStatusRetry(jobId);
-      } else if (normalizedStatus === 'pending') {
-        const meta = STATUS_RETRY_TIMERS.get(jobId);
-        if (!meta || meta.timer === null) {
-          scheduleStatusRetry(jobId, meta?.attempt ?? 0);
-        }
-      }
-    });
-    STATUS_RETRY_TIMERS.forEach((_meta, jobId) => {
-      if (!seen.has(jobId)) {
-        clearStatusRetry(jobId);
-      }
-    });
-  }, [swr.data]);
-
-  const stableJobs = stableStore.order.map((id) => stableStore.byId[id]).filter(Boolean);
-
-  return { ...swr, stableJobs } as typeof swr & { stableJobs: Job[] };
-}
-
-export async function getJobStatus(jobId: string): Promise<JobStatusResult> {
-  let response: Response;
-  try {
-    response = await authFetch(`/api/jobs/${encodeURIComponent(jobId)}`);
-  } catch (error) {
-    if (typeof window !== 'undefined') {
-      window.dispatchEvent(new CustomEvent('jobs:status-error', { detail: { jobId } }));
-    }
-    throw error;
-  }
-
-  const payload = (await response.json().catch(() => null)) as
-    | (Partial<JobStatusResult> & { error?: string; status?: string; progress?: number })
-    | null;
-
-  if (!response.ok) {
-    if (typeof window !== 'undefined') {
-      window.dispatchEvent(new CustomEvent('jobs:status-error', { detail: { jobId } }));
-    }
-    const message = payload?.error ?? `Status fetch failed (${response.status})`;
-    const error = new Error(message);
-    Object.assign(error, { status: response.status, jobId });
-    throw error;
-  }
-
-  if (!payload) {
-    if (typeof window !== 'undefined') {
-      window.dispatchEvent(new CustomEvent('jobs:status-error', { detail: { jobId } }));
-    }
-    throw new Error('Status payload missing');
-  }
-
-  const renderIds =
-    Array.isArray(payload.renderIds) && payload.renderIds.length
-      ? payload.renderIds.filter((value): value is string => typeof value === 'string')
-      : payload.renderIds === null
-        ? null
-        : undefined;
-  const renderThumbUrls =
-    Array.isArray(payload.renderThumbUrls) && payload.renderThumbUrls.length
-      ? payload.renderThumbUrls.filter((value): value is string => typeof value === 'string')
-      : payload.renderThumbUrls === null
-        ? null
-        : undefined;
-  const hasMedia = Boolean(payload.videoUrl || payload.audioUrl) || Boolean(renderIds?.length);
-  const statusFromPayload = normalizeJobStatus(payload.status ?? null, hasMedia);
-  const normalizedStatus = (statusFromPayload ?? (hasMedia ? 'completed' : 'pending')) as JobStatusResult['status'];
-  const progressValue = normalizeJobProgress(payload.progress, normalizedStatus, hasMedia);
-  const progress = progressValue ?? (normalizedStatus === 'completed' ? 100 : 0);
-  const normalizedMessage =
-    normalizeJobMessage(payload.message) ??
-    (normalizedStatus === 'failed'
-      ? 'The service reported a failure without details. Try again. If it fails repeatedly, contact support with your request ID.'
-      : undefined);
-
-  const result: JobStatusResult = {
-    ok: true,
-    jobId: payload.jobId ?? jobId,
-    status: normalizedStatus,
-    progress,
-    videoUrl: payload.videoUrl ?? null,
-    previewVideoUrl: payload.previewVideoUrl ?? null,
-    audioUrl: payload.audioUrl ?? null,
-    thumbUrl: payload.thumbUrl ?? null,
-    aspectRatio: payload.aspectRatio ?? null,
-    pricing: payload.pricing,
-    finalPriceCents: payload.finalPriceCents,
-    currency: payload.currency,
-    paymentStatus: payload.paymentStatus ?? 'platform',
-    batchId: payload.batchId ?? null,
-    groupId: payload.groupId ?? null,
-    iterationIndex: payload.iterationIndex ?? null,
-    iterationCount: payload.iterationCount ?? null,
-    renderIds: renderIds ?? null,
-    renderThumbUrls: renderThumbUrls ?? null,
-    heroRenderId: payload.heroRenderId ?? null,
-    localKey: payload.localKey ?? null,
-    message: normalizedMessage ?? null,
-    etaSeconds: payload.etaSeconds ?? null,
-    etaLabel: payload.etaLabel ?? null,
-  };
-
-  if (typeof window !== 'undefined') {
-    window.dispatchEvent(new CustomEvent<JobStatusResult>('jobs:status', { detail: result }));
-  }
-
-  return result;
-}
-
-type SavedAsset = {
-  id: string;
-  url: string;
-  width?: number | null;
-  height?: number | null;
-  mime?: string | null;
-  size?: number | null;
-};
-
-export async function saveAssetToLibrary(payload: {
-  url: string;
-  jobId?: string | null;
-  label?: string | null;
-  source?: string | null;
-  kind?: 'image' | 'video' | 'audio';
-  sourceOutputId?: string | null;
-  thumbUrl?: string | null;
-  previewUrl?: string | null;
-}) {
-  const response = await authFetch('/api/media-library/ensure', {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({
-      url: payload.url,
-      jobId: payload.jobId ?? null,
-      name: payload.label ?? null,
-      label: payload.label ?? null,
-      source: payload.source ?? null,
-      kind: payload.kind ?? 'image',
-      sourceOutputId: payload.sourceOutputId ?? null,
-      thumbUrl: payload.thumbUrl ?? null,
-      previewUrl: payload.previewUrl ?? null,
-    }),
-  });
-  const data = (await response.json().catch(() => null)) as { ok?: boolean; error?: string; asset?: SavedAsset } | null;
-  if (!response.ok || !data?.ok || !data.asset) {
-    throw new Error(
-      data?.error ??
-        `Failed to save ${payload.kind === 'video' ? 'video' : 'image'} (${response.status})`
-    );
-  }
-  return data.asset;
-}
-
-export async function saveImageToLibrary(payload: {
-  url: string;
-  jobId?: string | null;
-  label?: string | null;
-  source?: string | null;
-}) {
-  return saveAssetToLibrary({
-    ...payload,
-    kind: 'image',
-    source: payload.source ?? (payload.jobId ? 'generated' : null),
-  });
-}
-
-export async function hideJob(jobId: string): Promise<void> {
-  const response = await authFetch(`/api/jobs/${encodeURIComponent(jobId)}`, {
-    method: 'PATCH',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ hidden: true }),
-  });
-  const payload = (await response.json().catch(() => null)) as { ok?: boolean; error?: string } | null;
-  if (!response.ok || payload?.ok === false) {
-    throw new Error(payload?.error ?? `Failed to hide job (${response.status})`);
-  }
-  if (typeof window !== 'undefined') {
-    window.dispatchEvent(new CustomEvent('jobs:hidden', { detail: { jobId } }));
-  }
-}
+export { saveAssetToLibrary, saveImageToLibrary } from '@/lib/api-assets';
+export { useEngines } from '@/lib/api-engines';
+export { getJobStatus } from '@/lib/api-job-status';
+export { hideJob, useInfiniteJobs } from '@/lib/api-jobs';
